Delete the types module and bytes_read from Restic::ls
